detail

+++ app/Http/Controllers/QuestionsController.php
@@ -52,7 +52,9 @@ class QuestionsController extends Controller
      */
     public function show(Question $question)
     {
-        //
+        $question->increment('views');
+
+        return view('questions.show', compact('question'));
     }
 
     /**
@@ -63,6 +65,9 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Access denied");
+        }
         return view("questions.edit", compact('question'));
     }
 
@@ -75,6 +80,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Access denied");
+        }
         $question->update($request->only('title', 'body'));
 
         return redirect('/questions')->with('success', "Your question has been updated.");
@@ -88,6 +96,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Access denied");
+        }
         $question->delete();
 
         return redirect('/questions')->with('success', "Your question has been deleted.");
